Update GCD_algorithm.py
I added two functions;
Is_float(num): which checks if the number is a float

User_input() takes input from the user and ensures that it is valid according to the conditions of the Euclidean function, which means that float and negative numbers are not allowed, and both numbers can't be zero.

I have also added comments to my whole code explaining everything in detail.